Client.Common - move nes encode/decoder to a cheats folder
